Add basic mobile layout, desktop layout
After playing with this thing to try to come up with a "one size fits
all" solution for table column sizing, it occured to me that it may be
more productive to think of things in terms of the actual data in each
column, as opposed to trying to automatically size the columns for
arbitrary content.

So, that's where we are! On desktop columns are sized according to
reasonable min/max character values, and mobile is sized for the custom
mobile layout.